UBUNTU SAUCE: ubuntu stress smoke test: settings tweaks for slow RISC-V systems
On really slow devices the stress smoke test can take a long time to build
and run to completion.  Twiddle the settings to cater for these systems
by doubling the timeout time to 1hr, reducing the minimum free disk. Also
don't run with cgexec memory constraints if the system can't or won't
support this.

Signed-off-by: Colin Ian King <colin.king@canonical.com>