UBUNTU: SAUCE: ubuntu_ltp_syscalls: change the way how we print the SHA1
Since it's better to use a shallow clone to save us some time and
reduce the load to our server. We will have to change the way how we
print the SHA1 for debugging prupose.

As I always put the upstream SHA1 into the title of our LTP fork while
handling rebases and local patches, I think it's enough to just print
out the first commit title. Just need to remember to do this when
dealing with the LTP repo and we will be good.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>