UBUNTU: SAUCE: ubuntu_ltp*: print test summary in the end
If you click on a failed test case on the new result hinting system,
it will bring you to the bottom of the test.

So it's better to print out the summary section in the end of each
test instead of the beginning to make it easier for reviewers to
know which one is actually failing.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>