UBUNTU: SAUCE: ubuntu_lxc: stop and destroy the leftover reboot container
BugLink: https://bugs.launchpad.net/bugs/1941063

The leftover reboot container caused by bug 1776381 will prevent the
whole ubuntu_lxc test to finish properly on s390x instances, you will
see the test timeout in the end on the jenkins and leave the overall
result in an Incomplete state 'I' on the result page.

Solve this by stopping and destroying the reboot container in the end
of this test suite.

Note that the container must be stopped otherwise it will complain
about it's still running.

I have also removed the stdout / stderr redirection for the attempt
to destry the reboot container before the test to get a picture of
why this is still failing with the Container "reboot" is defined
error (bug 1943735). Normally this container will be stopped and can
be removed after rebooting the system, so I am not sure why this is
still happening for the moment.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>