UBUNTU: SAUCE: ubuntu_ltp_controllers: split controllers test out of ubuntu_ltp
Move the controller test out of the LTP test suite, and run it like
how we did for the ubuntu_ltp_syscalls test. This change can improve
our test granularity and make the result hinting become more precise.

The shortcoming will be the cost of the extra deployment time, test
build time and the extra storage space it will take.

This patch works as expected on our dev jenkins with a Hirsute KVM
instance.

Note that the test plan in CKCT needs to be updated as well.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>