UBUNTU: SAUCE: ubuntu_seccomp: build test cases in setup()
We will run `make` to build the cloned source code, but it will not
compile those test cases we need. So when the `make check` was
executed later, it will start building those test cases.

Use `make check-build` to build the library and all tests in setup().
This can make the test log become more easier to understand.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>