UBUNTU: SAUCE: ubuntu_ltp_kernel_misc: split kernel_misc test out of ubuntu_ltp
Move the kernel_misc test out of the LTP test suite, and run it like
how we did for the ubuntu_ltp_syscalls test. This change can improve
our test granularity and make the result hinting become more precise.

The ubuntu_ltp test on KVM kernels is hanging with the tpci test in
this kernel_misc test (lp:1947422), it should be able to finish
properly with this test split out. (But it is still expected to see
it hanging with this ubuntu_ltp_kernel_misc test.)

The downside of this is that it will cost some extra time for the
deployment and building the test suite.

This change will require a test plan update on CKCT as well.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>