UBUNTU: SAUCE: ubuntu_dgx_mofed_build: create build sanity test for MOFED DKMS on DGX
The goal of the test is to detect when a kernel SRU contains changes
that would cause the MOFED DKMS packages to fail to build. The test will
check if MOFED DKMS are built and loaded correctly on target platforms
currently. This test could help us to catch silent failure of DKMS build
which we have seen at least once.

The MOFED versions selected are limited to the versions recommended for
use with DGX systems.

The test environment is mostly prepared by MAAS via a customized curtin
preseed. For tasks like driver installation, software installation
highly associated with driver or required reboot, are setup by the
preseed. The rest of tasks are completed by the autotest framework, and
defined in the corresponding testing job.

Signed-off-by: Taihsiang Ho (tai271828) <taihsiang.ho@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>