UBUNTU: SAUCE: Use a more generic if any() statement to install modules-package
Use any() to match the flavour string of different clouds. So that we
don't need to keep adding new flavours to the list for new cloud
variants in the future.

Checked with "grep -r modules-extra * -B2" to see which test case needs
to be modified.

Note that not every test case needs the trusty AWS filter, as they
have been blacklisted with the blacklist file already.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>