UBUNTU: SAUCE: ubuntu_kernel_selftests: fix ip6tables missing issue
BugLink: https://bugs.launchpad.net/bugs/1952247

Add iptables package into the required package list to fix the
ip6tables missing issue in ip_defrag.sh test:
  # exec of "ip6tables" failed: No such file or directory

This issue was spotted on a manually provisioned Jammy RISCV
instance only, so I think it has something to do with our setup
or the package dependency of a RISCV instance. It's harmless to
add this package dependency, This package is available since Bionic.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>