UBUNTU: SAUCE: ubuntu_kernel_selftests: fix ethtool missing issue
On a Jammy RISCV instance, the net:udpgro_fwd.sh test is complaining about:
    exec of ethtool failed: No such file or directory

This issue was spotted on a manually provisioned Jammy RISCV
instance only, so I think it has something to do with our setup
or the package dependency of a RISCV instance. It's harmless to
add this package dependency, it is available since Bionic.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>