UBUNTU: SAUCE: ubuntu_nbd_smoke_test: wait before removing the module
BugLink: https://bugs.launchpad.net/bugs/1814572

Investigation on AWS a1.medium shows this module is not being used
after this test and can be removed, it's probably because the server
was not fully stopped yet when we try to unload the module.

Add a sleep to allow it to spin down.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>